Add files that will eventually be modified
Problem: Currently there are no well-personalized configuration files for development tools that are used quite often, and this does not make a lot of sense. There was also not much (or any, really) organization about which tools should be modified at all.

Solution: Create folders representing some of the initial tools that should be configured and personalized.

Testing: Initial configurations are simply defaults, not even modified from the local device from which they were pulled.

Signed-off-by: shaydonb