Add setup script and README
Problem: There was not yet a way for the configuration files to be easily integrated with the local environment from this repository. There also was not yet a README describing what this repository is.

Solution: Add a script that, although does not have any features in place yet, describes many of the ideated features in order to stay organized when working on this project in the future.

Testing: None necessary, no code yet in the script and just markdown for the README.md

Signed-off-by: shaydonb